Move get comicinfo & wails directory code to separate file
